`border-collapse: collapse` and `position: sticky` don't mix
`border-collapse: collapse` and `position: sticky` (or maybe
`position` in general) are not compatible – the collapsed borders do
not reappear when the cells are being positioned, and they arguably
should. But the specs say nothing about this and browsers do not
implement any such thing.

If we have sticky headers, then we shouldn't use `border-collapse:
collapse`. It's easy enough to do that, just ensure the table cells
only have borders on one side, it's just a little tedious.

This patch does that, and removes the `box-shadow` hack, which looked
ugly.