Use the new MobileFrontendContentProvider
Add the extension to the list but only enable it if proxy
is set or its added in the dropdown.

Drop configuration variables as these are all the defaults
in the new extension.

Should not be merged until https://gerrit.wikimedia.org/r/c/mediawiki/extensions/MobileFrontend/+/709544
has been merged.