Add ElasticSearch service
* Install Docker in VM creation
* Run ElasticSearch as a container using the image from
releng/dev-images which has WMF plugins
* Run CirrusSearch setup scripts in postinstall.sh

Periodically cleaning up old search index data is not done in this
patch. It can be done by deleting the volume associated with the
container (since this is considered throwaway data).

It would be nice if runJobs.php ran periodically for the wiki, as that's
needed to process updates to the search index. Ideas welcome on how to
do that.

Fixes #356